microservices/customer/pyproject.toml
[tool.poetry]
name = "customer"
version = "0.1.0"
description = ""
authors = ["Clariteia Devs <devs@clariteia.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
minos-microservice-common = "^0.5.0"
minos-microservice-networks = "^0.5.0"
minos-microservice-aggregate = "^0.5.0"
minos-microservice-saga = "^0.5.0"
minos-microservice-cqrs = "^0.5.0"
minos-broker-kafka = "^0.5.0"
minos-discovery-minos = "^0.5.0"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^19.10b"
isort = "^5.8.0"
flake8 = "^3.9.2"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
microservice = "src.cli:main"